update schema create/destroy

+++ store/schema.js
@@ -5,28 +5,36 @@ Create the schema for our database
 */
 exports.create = function(knex) {
   return knex.transaction(function(t) {
-    var Tasks = knex.schema.createTable('tasks', function(t) {
-      t.uuid('taskId').primary();
-      t.string('provisionerId', 22).notNullable();
-      t.string('workerType', 22).notNullable();
-      t.string('state').notNullable();
-      t.string('reason').notNullable();
-      t.string('routing', 128).notNullable();
-      t.integer('retries').notNullable();
-      t.integer('timeout').notNullable();
-      t.specificType('priority', 'double precision').notNullable();
-      t.timestamp('created').notNullable();
-      t.timestamp('deadline').notNullable();
-      t.timestamp('takenUntil').notNullable();
-    }).transacting(t);
+    var Tasks = knex.schema.hasTable('tasks').
+      then(function(exists) {
+        if (exists) return;
+        return knex.schema.createTable('tasks', function(t) {
+          t.uuid('taskId').primary();
+          t.string('provisionerId', 22).notNullable();
+          t.string('workerType', 22).notNullable();
+          t.string('state').notNullable();
+          t.string('reason').notNullable();
+          t.string('routing', 128).notNullable();
+          t.integer('retries').notNullable();
+          t.integer('timeout').notNullable();
+          t.specificType('priority', 'double precision').notNullable();
+          t.timestamp('created').notNullable();
+          t.timestamp('deadline').notNullable();
+          t.timestamp('takenUntil').notNullable();
+        }).transacting(t);
+      });
 
-    var Runs = knex.schema.createTable('runs', function(t) {
-      t.uuid('taskId').references('taskId').inTable('tasks').onDelete('cascade');
-      t.integer('runId').notNullable();
-      t.string('workerGroup', 22).notNullable();
-      t.string('workerId', 22).notNullable();
-      t.primary(['taskId', 'workerId']);
-    }).transacting(t);
+    var Runs = knex.schema.hasTable('runs').
+      then(function(exists) {
+        if (exists) return;
+        return knex.schema.createTable('runs', function(t) {
+          t.uuid('taskId').references('taskId').inTable('tasks').onDelete('cascade');
+          t.integer('runId').notNullable();
+          t.string('workerGroup', 22).notNullable();
+          t.string('workerId', 22).notNullable();
+          t.primary(['taskId', 'workerId']);
+        }).transacting(t);
+      });
 
     return Promise.all([Tasks, Runs]).then(t.commit, t.rollback);
   });
